requirement.txt
google-generativai
